fix(migration): product_color_prices i32 列 created_by/approved_by/customer_id BIGINT→INTEGER
模型 61a7150d 已改 Option<i32>(引用 users.id/customers.id SERIAL i32),但 sales_crm 建列仍 BIGINT;
wave5c global seed 首插该表→SeaORM RETURNING 读回 INT8 解码 i32 失败→#4647 traversal(4/5)+flow(9)
color-card 500。该表无 TYPE BIGINT 加宽路径(加宽仅 custom_orders),且 m0044 同表这三列本就用 INTEGER,
方向无歧义。全新库 ADD COLUMN IF NOT EXISTS 首建零漂移。未碰 custom_orders 家族。

+++ backend/migration/src/domain/sales_crm/mod.rs
@@ -388,13 +388,13 @@ SELECT setval('crm_tag_id_seq', COALESCE((SELECT MAX(id) FROM "crm_tag"), 0) + 1
 
 
 -- === 从旧迁移恢复的 ALTER ADD COLUMN（确保迁移表结构与 Model 一致）===
-ALTER TABLE "product_color_prices" ADD COLUMN IF NOT EXISTS "approved_by" BIGINT;
+ALTER TABLE "product_color_prices" ADD COLUMN IF NOT EXISTS "approved_by" INTEGER;
 ALTER TABLE "product_color_prices" ADD COLUMN IF NOT EXISTS "max_quantity" DECIMAL(18,2);
 ALTER TABLE "product_color_prices" ADD COLUMN IF NOT EXISTS "season" VARCHAR(10);
-ALTER TABLE "product_color_prices" ADD COLUMN IF NOT EXISTS "created_by" BIGINT;
+ALTER TABLE "product_color_prices" ADD COLUMN IF NOT EXISTS "created_by" INTEGER;
 ALTER TABLE "product_color_prices" ADD COLUMN IF NOT EXISTS "is_active" BOOLEAN;
 ALTER TABLE "product_color_prices" ADD COLUMN IF NOT EXISTS "priority" INT;
-ALTER TABLE "product_color_prices" ADD COLUMN IF NOT EXISTS "customer_id" BIGINT;
+ALTER TABLE "product_color_prices" ADD COLUMN IF NOT EXISTS "customer_id" INTEGER;
 ALTER TABLE "product_color_prices" ADD COLUMN IF NOT EXISTS "approval_status" VARCHAR(20);
 ALTER TABLE "product_color_prices" ADD COLUMN IF NOT EXISTS "approved_at" TIMESTAMPTZ;
 ALTER TABLE "sales_quotations" ADD COLUMN IF NOT EXISTS "freight_cost" DECIMAL(14,2);
